Blur benchmark + difference scorer using Bild

+++ src/changeDetection/changeDetection.go
@@ -4,5 +4,5 @@ import "mjpeg_multiplexer/src/mjpeg"
 
 // ChangeDetectionScorer compares multiple images and emits difference as int score
 type ChangeDetectionScorer interface {
-	Score([]mjpeg.MjpegFrame) int
+	Score([]mjpeg.MjpegFrame) float64
 }
